things print
