+++ src/App/App.js
@@ -3,16 +3,20 @@ import './App.scss';
 
 import studentsData from '../helpers/data/studentsData';
 
-import SharkTank from '../components/components/ShartTank/SharkTank';
+import SharkTank from '../components/SharkTank/SharkTank';
+import Graveyard from '../components/Graveyard/Graveyard';
 
 class App extends React.Component {
   state = {
     students: [],
+    deadOnes: [],
   }
 
   componentDidMount() {
     const students = studentsData.livingStudents();
     this.setState({ students });
+    const deadOnes = studentsData.dearlyBeloved();
+    this.setState({ deadOnes });
   }
 
   render() {
@@ -20,7 +24,15 @@ class App extends React.Component {
       <div className="App">
         <h1>Shark Tank</h1>
           <button className='btn btn-danger'>SHARK ATTACK</button>
-          <SharkTank students={this.state.students} />
+          <div>
+            <SharkTank className='container' students={this.state.students} />
+          </div>
+          <div>
+            <h1>Graveyard</h1>
+              <div>
+                <Graveyard className='container' deadOnes={this.state.deadOnes} />
+              </div>
+          </div>
       </div>
     );
   }
