+++ src/components/GraveStone/GraveStone.js
@@ -1,8 +1,8 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import studentShape from '../../../helpers/propz/studentShade';
+import studentShape from '../../helpers/propz/studentShade';
 
-class LiveStudents extends React.Component {
+class GraveStone extends React.Component {
   static propTypes = {
     student: studentShape.studentShape,
   }
@@ -12,7 +12,7 @@ class LiveStudents extends React.Component {
     const { student } = this.props;
 
     return (
-      <div className="card col-3">
+      <div className="card col-6">
         <div className="card-body">
           <h5 className="card-title">{student.firstName} {student.lastName}</h5>
         </div>
@@ -21,4 +21,4 @@ class LiveStudents extends React.Component {
   }
 }
 
-export default LiveStudents;
+export default GraveStone;
