--- src/components/components/GraveStone/GraveStone.js
@@ -1,18 +1,0 @@
-import React from 'react';
-import PropTypes from 'prop-types';
-import studentShape from '../../../helpers/propz/studentShade';
-
-class GraveStone extends React.Component {
-  static propTypes = {
-    student: studentShape.studentShape,
-  }
-
-  render() {
-    const { student } = this.props;
-
-    return (
-      
-    )
-  }
-
-}
