--- src/components/components/Graveyard/Graveyard.js
@@ -1,1 +1,0 @@
-// import React from 'react';
