+++ src/styles/index.scss
@@ -8,10 +8,7 @@ body {
     sans-serif;
   -webkit-font-smoothing: antialiased;
   -moz-osx-font-smoothing: grayscale;
-  background-image: url("https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcQ9awmLWgE68GYFZ797h3Nzr7xG0d1l7Lfw-FQIX-NHrrvbTm95rw&s");
-  background-position: center;
-  background-repeat: no-repeat;
-  background-size: cover;
+  background-color: $green;
 }
 
 code {
